extract registry from icecast
